feat: enrich stats with cache/pricing/model info
- IBlockMemoryProvider: add GetUsageAsync
- BlockMemoryProvider: implement GetUsageAsync (delegates to store)
- MemoryTool: accept IConfigService for pricing, enrich stats output
  with model name, cache hit/miss rate, cumulative cost
- ToolFactory/ToolRegistry: pass cfgService to MemoryTool
- /stats command: add model, cumulative cache hit/miss/rate, cost
- ConsoleRenderer.FormatTokenCount: int→long (usage values are long)

+++ src/Glyphite.Cli/Services/ConsoleRenderer.cs
@@ -205,7 +205,7 @@ public class ConsoleRenderer
         }
     }
 
-    public static string FormatTokenCount(int count) => count >= 1000
+    public static string FormatTokenCount(long count) => count >= 1000
         ? $"{count / 1000.0:F1}K"
         : count.ToString();
 }
